# internal/docs/swagger.yaml
basePath: /
definitions:
...
  internal_social.followRequest:
    properties:
      vlogger_id:
        type: integer
    type: object
  internal_video.Comment:
    properties:
      author_id:
        type: integer
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      username:
        type: string
      video_id:
        type: integer
    type: object
  internal_video.Video:
    properties:
      author_id:
        type: integer
      cover_url:
        type: string
      create_time:
        type: string
      description:
        type: string
      id:
        type: integer
      likes_count:
        type: integer
      play_url:
        type: string
      popularity:
        type: integer
      title:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  internal_video.deleteCommentRequest:
    properties:
      comment_id:
        type: integer
    type: object
  internal_video.deleteVideoRequest:
    properties:
      id:
        type: integer
    type: object
  internal_video.getAllCommentsRequest:
    properties:
      video_id:
        type: integer
    type: object
  internal_video.getDetailRequest:
    properties:
      id:
        type: integer
    type: object
...
  internal_video.publishCommentRequest:
    properties:
      content:
        type: string
      video_id:
        type: integer
    type: object
  internal_video.publishRequest:
    properties:
      cover_url:
        type: string
      description:
        type: string
      play_url:
        type: string
      title:
        type: string
    type: object
host: 127.0.0.1:8080
info:
  contact: {}
  description: 短视频社区后端 API:账号、视频、Feed、互动、通知。
  title: Video Feed API
  version: "1.0"
paths:
...
  /comment/delete:
    post:
      consumes:
      - application/json
      parameters:
      - description: 评论 ID
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_video.deleteCommentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 删除评论
      tags:
      - comment
  /comment/getAll:
    post:
      consumes:
      - application/json
      parameters:
      - description: 视频 ID
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_video.getAllCommentsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_video.Comment'
            type: array
      summary: 获取评论列表
      tags:
      - comment
  /comment/publish:
    post:
      consumes:
      - application/json
      description: 通过 MQ 异步落库,失败降级同步写
      parameters:
      - description: 评论参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_video.publishCommentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_video.Comment'
      security:
      - BearerAuth: []
      summary: 发表评论
      tags:
      - comment
...
  /social/follow:
    post:
      consumes:
      - application/json
      description: 通过 MQ 异步落库,失败降级同步写
      parameters:
      - description: 目标用户 ID
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_social.followRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 关注用户
      tags:
      - social
  /social/unfollow:
    post:
      consumes:
      - application/json
      parameters:
      - description: 目标用户 ID
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_social.followRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 取消关注
      tags:
      - social
  /video/delete:
    post:
      consumes:
      - application/json
      description: 仅作者本人可删除
      parameters:
      - description: 视频 ID
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_video.deleteVideoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 删除视频
      tags:
      - video
  /video/getDetail:
    post:
      consumes:
      - application/json
      description: 本地缓存 → Redis MGET → MySQL 三级路径,singleflight 合并回源
      parameters:
      - description: 视频 ID
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_video.getDetailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_video.Video'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 视频详情
      tags:
      - video
  /video/publish:
    post:
      consumes:
      - application/json
      description: 创建视频记录,事务内同写 Outbox 由轮询器投递时间线事件
      parameters:
      - description: 发布参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_video.publishRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_video.Video'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 发布视频
      tags:
      - video
  /video/uploadVideo:
    post:
      consumes:
      - multipart/form-data
      description: multipart/form-data 上传 .mp4,返回 URL
      parameters:
      - description: 视频文件 (.mp4, ≤200MB)
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 上传视频文件
      tags:
      - video
securityDefinitions:
  BearerAuth:
    description: Bearer JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"