# internal/video/handler.go
package video

import (
	"crypto/rand"
	"encoding/hex"
	"errors"
	"fmt"
	"mime/multipart"
	"net/http"
	"os"
	"path"
	"path/filepath"
	"strings"
	"time"

	authjwt "video-feed/internal/middleware/jwt"

	"github.com/gin-gonic/gin"
	"gorm.io/gorm"
)

type Handler struct {
	service *Service
}

func NewHandler(service *Service) *Handler {
	return &Handler{service: service}
}

// Publish godoc
// @Summary      发布视频
// @Description  创建视频记录,事务内同写 Outbox 由轮询器投递时间线事件
// @Tags         video
// @Accept       json
// @Produce      json
// @Security     BearerAuth
// @Param        body  body      publishRequest  true  "发布参数"
// @Success      200   {object}  Video
// @Failure      400   {object}  map[string]string
// @Failure      401   {object}  map[string]string
// @Router       /video/publish [post]
func (h *Handler) Publish(c *gin.Context) {
	var req publishRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}
	accountID, err := authjwt.GetAccountID(c)
	if err != nil {
		c.JSON(http.StatusUnauthorized, gin.H{"error": err.Error()})
		return
	}
	username, err := authjwt.GetUsername(c)
	if err != nil {
		c.JSON(http.StatusUnauthorized, gin.H{"error": err.Error()})
		return
	}

	video, err := h.service.Publish(c.Request.Context(), PublishInput{
		AuthorID:    accountID,
		Username:    username,
		Title:       req.Title,
		Description: req.Description,
		PlayURL:     req.PlayURL,
		CoverURL:    req.CoverURL,
	})
	if err != nil {
		c.JSON(statusForVideoError(err), gin.H{"error": err.Error()})
		return
	}

	c.JSON(http.StatusOK, video)
}

// UploadVideo godoc
// @Summary      上传视频文件
// @Description  multipart/form-data 上传 .mp4,返回 URL
// @Tags         video
// @Accept       multipart/form-data
// @Produce      json
// @Security     BearerAuth
// @Param        file  formData  file  true  "视频文件 (.mp4, ≤200MB)"
// @Success      200   {object}  map[string]string
// @Router       /video/uploadVideo [post]
func (h *Handler) UploadVideo(c *gin.Context) {
	accountID, err := authjwt.GetAccountID(c)
	if err != nil {
		c.JSON(http.StatusUnauthorized, gin.H{"error": err.Error()})
		return
	}

	file, err := c.FormFile("file")
	if err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": "missing file"})
		return
	}

	const maxVideoSize = 200 << 20
	if file.Size <= 0 || file.Size > maxVideoSize {
		c.JSON(http.StatusBadRequest, gin.H{"error": "invalid file size"})
		return
	}

	ext := strings.ToLower(filepath.Ext(file.Filename))
	if ext != ".mp4" {
		c.JSON(http.StatusBadRequest, gin.H{"error": "only .mp4 is allowed"})
		return
	}

	url, err := h.saveUploadedFile(c, file, "videos", accountID, ext)
	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
		return
	}
	c.JSON(http.StatusOK, gin.H{"url": url, "play_url": url})
}

func (h *Handler) UploadCover(c *gin.Context) {
	accountID, err := authjwt.GetAccountID(c)
	if err != nil {
		c.JSON(http.StatusUnauthorized, gin.H{"error": err.Error()})
		return
	}

	file, err := c.FormFile("file")
	if err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": "missing file"})
		return
	}

	const maxCoverSize = 10 << 20
	if file.Size <= 0 || file.Size > maxCoverSize {
		c.JSON(http.StatusBadRequest, gin.H{"error": "invalid file size"})
		return
	}

	ext := strings.ToLower(filepath.Ext(file.Filename))
	switch ext {
	case ".jpg", ".jpeg", ".png", ".webp":
	default:
		c.JSON(http.StatusBadRequest, gin.H{"error": "only .jpg/.jpeg/.png/.webp is allowed"})
		return
	}

	url, err := h.saveUploadedFile(c, file, "covers", accountID, ext)
	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
		return
	}
	c.JSON(http.StatusOK, gin.H{"url": url, "cover_url": url})
}

// DeleteVideo godoc
// @Summary      删除视频
// @Description  仅作者本人可删除
// @Tags         video
// @Accept       json
// @Produce      json
// @Security     BearerAuth
// @Param        body  body      deleteVideoRequest  true  "视频 ID"
// @Success      200   {object}  map[string]string
// @Failure      401   {object}  map[string]string
// @Failure      404   {object}  map[string]string
// @Router       /video/delete [post]
func (h *Handler) DeleteVideo(c *gin.Context) {
	var req deleteVideoRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}
	accountID, err := authjwt.GetAccountID(c)
	if err != nil {
		c.JSON(http.StatusUnauthorized, gin.H{"error": err.Error()})
		return
	}
	if err := h.service.Delete(c.Request.Context(), req.ID, accountID); err != nil {
		c.JSON(statusForVideoError(err), gin.H{"error": err.Error()})
		return
	}
	c.JSON(http.StatusOK, gin.H{"message": "video deleted"})
}

func (h *Handler) saveUploadedFile(c *gin.Context, file *multipart.FileHeader, category string, accountID uint, ext string) (string, error) {
	date := time.Now().Format("20060102")
	relDir := filepath.Join(category, fmt.Sprintf("%d", accountID), date)
	root := filepath.Join(".run", "uploads")
	absDir := filepath.Join(root, relDir)
	if err := os.MkdirAll(absDir, 0o755); err != nil {
		return "", err
	}

	name, err := randHex(16)
	if err != nil {
		return "", err
	}
	filename := name + ext
	absPath := filepath.Join(absDir, filename)
	if err := c.SaveUploadedFile(file, absPath); err != nil {
		return "", err
	}

	urlPath := path.Join("/static", category, fmt.Sprintf("%d", accountID), date, filename)
	return buildAbsoluteURL(c, urlPath), nil
}

func randHex(size int) (string, error) {
	buf := make([]byte, size)
	if _, err := rand.Read(buf); err != nil {
		return "", fmt.Errorf("generate random filename: %w", err)
	}
	return hex.EncodeToString(buf), nil
}

func buildAbsoluteURL(c *gin.Context, urlPath string) string {
	scheme := "http"
	if c.Request.TLS != nil {
		scheme = "https"
	}
	if forwarded := c.GetHeader("X-Forwarded-Proto"); forwarded != "" {
		scheme = forwarded
	}
	return fmt.Sprintf("%s://%s%s", scheme, c.Request.Host, urlPath)
}

// GetDetail godoc
// @Summary      视频详情
// @Description  本地缓存 → Redis MGET → MySQL 三级路径,singleflight 合并回源
// @Tags         video
// @Accept       json
// @Produce      json
// @Param        body  body      getDetailRequest  true  "视频 ID"
// @Success      200   {object}  Video
// @Failure      404   {object}  map[string]string
// @Router       /video/getDetail [post]
func (h *Handler) GetDetail(c *gin.Context) {
	var req getDetailRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	video, err := h.service.GetDetail(c.Request.Context(), req.ID)
	if err != nil {
		status := http.StatusInternalServerError
		if errors.Is(err, gorm.ErrRecordNotFound) {
			status = http.StatusNotFound
		}
		c.JSON(status, gin.H{"error": err.Error()})
		return
	}

	c.JSON(http.StatusOK, video)
}

func (h *Handler) ListByAuthorID(c *gin.Context) {
	var req listByAuthorIDRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	videos, err := h.service.ListByAuthorID(c.Request.Context(), req.AuthorID)
	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
		return
	}
	if videos == nil {
		videos = []Video{}
	}

	c.JSON(http.StatusOK, videos)
}

func statusForVideoError(err error) int {
	switch {
	case errors.Is(err, ErrInvalidInput):
		return http.StatusBadRequest
	case errors.Is(err, ErrUnauthorized):
		return http.StatusUnauthorized
	case errors.Is(err, gorm.ErrRecordNotFound):
		return http.StatusNotFound
	default:
		return http.StatusInternalServerError
	}
}